+++ AndroidProject/DeLighted/app/src/main/res/layout/tab1_frag.xml
@@ -1,0 +1,85 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".TabOneFragment"
+    android:orientation="vertical">
+
+    <LinearLayout
+        android:id="@+id/OutPutLayout"
+        android:layout_width="match_parent"
+        android:layout_height="141dp"
+        android:orientation="horizontal">
+
+        <EditText
+            android:id="@+id/OutPutText"
+            android:layout_width="0dp"
+            android:layout_height="136dp"
+            android:layout_weight="1"
+            android:ems="10"
+            android:inputType="textMultiLine"
+            android:text="text" />
+    </LinearLayout>
+
+    <LinearLayout
+        android:id="@+id/RequestButtons"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+        <Button
+            android:id="@+id/pick_date"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Set Date" />
+
+        <Button
+            android:id="@+id/pick_duration"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Set Duration" />
+
+        <Button
+            android:id="@+id/list_alarms"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="List Alarms" />
+
+        <Button
+            android:id="@+id/clear_alarms"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Clear Alarms" />
+
+        <Button
+            android:id="@+id/delete_alarm"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Delete" />
+    </LinearLayout>
+
+    <LinearLayout
+        android:id="@+id/URLConnection"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+
+        <EditText
+            android:id="@+id/URLText"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:ems="10"
+            android:inputType="textPersonName"
+            android:text="URL" />
+
+        <Button
+            android:id="@+id/SetURLButton"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="SetURL" />
+    </LinearLayout>
+
+</LinearLayout>
